remove query from fragment
