+++ packages/graphql/src/fragments/hubs.ts
@@ -117,29 +117,3 @@ export const HubAssetFragment = graphql(
   ],
 );
 export type HubAsset = FragmentOf<typeof HubAssetFragment>;
-
-/**
- * @internal
- */
-export const HubQuery = graphql(
-  `query Hub($request: HubRequest!) {
-      value: hub(request: $request) {
-        ...Hub
-      }
-    }`,
-  [HubFragment],
-);
-export type HubRequest = RequestOf<typeof HubQuery>;
-
-/**
- * @internal
- */
-export const HubsQuery = graphql(
-  `query Hubs($request: HubsRequest!) {
-      value: hubs(request: $request) {
-        ...Hub
-      }
-    }`,
-  [HubFragment],
-);
-export type HubsRequest = RequestOf<typeof HubsQuery>;
