+++ packages/graphql/src/index.ts
@@ -2,6 +2,7 @@ export * from './common';
 export * from './enums';
 export * from './fragments';
 export * from './graphql';
+export * from './hubs';
 export * from './misc';
 export * from './permits';
 export * from './transactions';
